1) Fixed 'moving physics', i.e. objects on the track that    can be pushed around by karts (currently only used in    tux tollway). 2) Removed support for magnet. 3) Some code cleanup for missiles etc.
git-svn-id: svn+ssh://svn.code.sf.net/p/supertuxkart/code/trunk/supertuxkart@1379 178a84e3-b1eb-0310-8ba1-8eac791a3b58

+++ src/widget.cpp
@@ -148,8 +148,8 @@ void Widget::update(const float DELTA)
 
         if( m_track_num != -1 )
         {
-            track_manager->getTrack( m_track_num )->drawScaled2D( m_x, m_y,
-                m_width, m_height);
+            track_manager->getTrack( m_track_num )->drawScaled2D( (float)m_x, 
+                (float)m_y, (float)m_width, (float)m_height);
         }
         else
         {
